commit-1-9-add-artistrepo

+++ src/main/java/ru/itis/repositories/CrudRepository.java
@@ -4,7 +4,7 @@ import java.util.List;
 import java.util.Optional;
 
 public interface CrudRepository<T> {
-    Optional<T> findOne(int id);
+    Optional<T> findOne(Long id);
     void save(T model);
     void delete(Long id);
     Optional<List<T>> findAll();
